integration searchBar - add font - refractor
