+++ client/src/App/assets/fonts/lato/Lato-HairlineItalic.ttf
