+++ client/src/App/assets/fonts/lato/Lato-Regular.ttf
