+++ client/src/App/components/breadcrumb/breadcrumb.scss
@@ -1,20 +1,20 @@
 @import '../../styles/variables';
 
 .breadcrumb {
-    text-decoration: none;
-    color: inherit;
-    color: $blue;
-    font-family: $font-bold;
-}
-
-.breadcrumb:hover {
-    text-decoration: underline;
-}
-
-.chevron:last-child {
-    display: none;
-}
-
-.chevron {
-    margin: 0 5px;
+    a {
+        font-family: $font-bold;
+        font-size: 12px;
+        color: $blue;
+        text-decoration: none;
+        &:hover {
+            text-decoration: underline;
+        }
+        &:not(:last-child) {
+            &:after {
+                content: '>';
+                display: inline-block;
+                margin: 0 5px;
+            }
+        }
+    }
 }
