+++ client/src/App/components/breadcrumb/index.jsx
@@ -1,20 +1,16 @@
-import React, {Fragment} from 'react'
+import React from 'react';
 import withBreadcrumbs from 'react-router-breadcrumbs-hoc';
 import {Link} from 'react-router-dom';
 
-import './breadcrumb.scss'
+import './breadcrumb.scss';
+
 const Breadcrumbs = ({ breadcrumbs }) => (
-  <div>
+  <div className="breadcrumb">
     {breadcrumbs.map(({
       match,
       breadcrumb
     }) => (
-    <Fragment>
-        <span key={match.url}>
-            <Link className="breadcrumb" to={match.url}>{breadcrumb}</Link> 
-        </span>
-        <span className="breadcrumb chevron">></span>
-    </Fragment>
+      <Link key={match.url} to={match.url}>{breadcrumb}</Link>
     ))}
   </div>
 );
