+++ client/src/App/components/champbubble/champbubble.scss
@@ -7,7 +7,6 @@
     background-size: cover;
     background-repeat: no-repeat;
     background-position: center center;
-    transition: all 0.3s ease;
 
     &:active,
     &:hover {
