+++ client/src/App/components/header/index.jsx
@@ -1,5 +1,7 @@
 import React, { Component } from 'react';
-import {Link} from 'react-router-dom';
+
+import Breadcrumbs from '../breadcrumb';
+import SearchBar from '../searchBar';
 
 import './header.scss';
 
@@ -8,16 +10,19 @@ import Logo from '../../assets/Logo.png';
 class Header extends Component {
     render() {
         return (
-            <header className="header wrapper">
+            <header className="header">
                 <div className="left">
                     <div className="logo">
                         <img src={Logo} alt="Logo" />
                     </div>
-                    <div className="breadcrumb">
-                        <Link to='#'>Home</Link>
-                    </div>
+                    <Breadcrumbs />
                 </div>
-                <div className="right icon icon-search"></div>
+                <div className="navigation">
+
+                </div>
+                <div className="right">
+                    <SearchBar /> 
+                </div>
             </header>
         )
     }
