+++ client/src/App/components/infoModal/infoModal.scss
@@ -20,12 +20,16 @@
     }
 }
 
-.modal-info {
+.modal {
     display: none;
     z-index: 100;
     position: fixed;
     height: 100vh;
     width: 100vw;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
     background: $darkblue;
     animation-name: animationGo;
     animation-duration: 0.3s;  
@@ -48,18 +52,18 @@
     max-width: 540px;
     text-align: center;
     .title {
-        margin: 0 112px;
+        margin: 0 100px;
         padding-bottom: 9px;
-        font-family: $font-bold;
-        font-size: 24px;
-        line-height: 36px;
+        font-family: $title;
+        font-size: 28px;
+        line-height: 38px;
         background: -webkit-linear-gradient(#C79A3C, #A6843C, #614B1D);
         -webkit-background-clip: text;
         -webkit-text-fill-color: transparent;
         border-bottom: solid 1px $gold;
     }
     .img-graph {
-        margin: 40px 0;
+        margin: 30px 0;
     }
     .img-rank {
         width: 274px;
@@ -75,17 +79,16 @@
         }
     }
     .btn-close {
-        margin-top: 44px;
+        margin-top: 30px;
         display: inline-block;
         padding: 16px;
         text-align: center;
         background: $blue;
         color: $darkblue;
-        font-size: 24px;
+        font-size: 18px;
         text-decoration: none;
         box-shadow: $box-shadow;
         user-select: none;
-        transition: background 0.3s ease;
         &:hover {
             background: $gold-gradient;
         }
@@ -93,6 +96,6 @@
 }
 
 @keyframes animationGo {
-    from { opacity: 0; transform: scale(0.7); }
-    to { opacity: 1; transform: scale(1); }
+    from { opacity: 0; transform: translateY(10px); }
+    to { opacity: 1; transform: translateY(0px); }
 }
