+++ client/src/App/components/postbubble/postbubble.scss
@@ -9,7 +9,6 @@
     height: 80px;
     border-radius: 50%;
     box-shadow: $box-shadow;
-    transition: all 0.3s ease;
     cursor: pointer;
     &.icon:before {
         top: 34%;
