+++ client/src/App/components/searchBar/searchBar.scss
@@ -1,0 +1,103 @@
+@import '../../styles/variables';
+
+.searchbar {
+    position: relative;
+    &:before {
+        position: absolute;
+        content: '\E801';
+        top: 50%;
+        right: 0;
+        transform: translateY(-50%);
+        font-family: 'fontello';
+        font-size: 32px;
+        color: $blue;
+    }
+    input {
+        width: 100%;
+        border: 0;
+        border-bottom: solid 1px $gold;
+        padding-bottom: 5px;
+        font-family: $font;
+        font-size: 16px;
+        color: $white;
+        outline: none;
+        background: none;
+        text-transform: capitalize;
+        &::placeholder {
+            font-family: $font;
+            font-size: 16px;
+            color: $blue;
+        }
+    }
+    input[type="search"]::-webkit-search-cancel-button {
+        display: none;
+    }
+    ~ ul {
+        margin: 18px auto 0;
+        width: 236px;
+        li {
+            animation: appear 0.3s ease-in;
+            &:hover {
+                .btn-champ {
+                    border-color: $gold;
+                }
+                a {
+                    background: -webkit-linear-gradient(#C79A3C, #A6843C, #614B1D);
+                    -webkit-background-clip: text;
+                    -webkit-text-fill-color: transparent;
+                }
+            }
+            &:not(:last-of-type) {
+                margin-bottom: 18px;
+            }
+            .btn-champ {
+                margin-right: 22px;
+            }
+            a {
+                display: flex;
+                justify-content: flex-start;
+                align-items: center;
+                font-family: $font;
+                font-size: 24px;
+                color: $blue;
+                text-decoration: none;
+            }
+        }
+        li.input-empty {
+            display: flex;
+            justify-content: flex-start;
+            align-items: center;
+            &:first-of-type {
+                opacity: 0.15;
+            }
+            &:nth-child(2) {
+                opacity: 0.1;
+            }
+            &:nth-child(3) {
+                opacity: 0.05;
+            }
+            .imgChamp {
+                flex-shrink: 0;
+                height: 56px;
+                width: 56px;
+                margin-right: 16px;
+                border-radius: 50%;
+                background: $gold-gradient;
+            }
+            .nameChamp {
+                width: 100%;
+                height: 28px;
+                background: $gold-gradient;
+            }
+        }
+    }
+}
+
+@keyframes appear {
+    from {
+        transform: translateY(4px);
+    }
+    to {
+        transform: translateY(0px);
+    }
+}
