+++ client/src/App/styles/_variables.scss
@@ -11,3 +11,5 @@ $box-shadow: 0px 0px 30px rgba(0, 203, 224, 0.15);
 
 $font: 'Lato';
 $font-bold: 'Lato-bold';
+$title: 'Martel';
+$title-bold: 'Martel-bold'
