--- client/src/App/views/landing/landing.scss
@@ -1,1 +1,0 @@
-@import '../../styles/variables';
