+++ client/src/data.json
@@ -11,7 +11,7 @@
         "control": "60",
         "mobility": "20",
         "utility": "10",
-        "img": "img",
+        "img": "http://elohell.net/public/champions/avatar/VelKozSquare1.png",
         "icon": "icon",
         "poste": {
             "mid": 38,
